Start to migrate over to node.js
The current plan is to use a central library (example_core.js) for
all the server-ing action. That is, handling incoming requests,
routing them and grabbing POST data and all that boring stuff that
very much depends on the language/framework/platform/whatever. What
you are left with is code that is fairly simple to understand and
that shows you what exactly is going on in a bit of an overview.

This commit includes that initial core library and a rewrite of
example number two - group messaging.