Reorganize the Node core library
The idea of having a nice function as an export was nice, but it
will probably make code less readable without knowing what is going
on. As the goal of the library was to abstract away all the backend
stuff, that is not good at all. With this commit, the core library
will abort right away on a bad core.listen() call instead of leaving
that up the the example which is probably for the best anyway.