Improve Node core library and add example one
The core library now adds charset=utf-8 to the content type of all
responses and provides the listeners with the response object so that
they can send their own data as a response as well. This is required
for the first example, which is now added as a node.js executable as
well.