Abstract away the response writing in Node
The Node core library now provides abstraction for response writing
as well. This is a fairly sensible move to maintain support for the
multi-listener system that was established and provide a content
length header at the same time. In previous testing we already
established that the endpoints can get a bit mad when that header
is missing so it's probably a good idea to include it here as well.