Port exercise four over to Node
This one is a bit special and includes changes to the core library.
Long story short, not all requests will necessarily be POST requests
and if we left it expecting a POST request, people would probably
be left asking why the example code is not working, which would give
a bad first impression.