Allow future expansion of Node core library
We are switching over to a single 'page' object for the listener
callbacks. As of right now that object comes with a 'post' object
that has the POST key-value pairs extracted and a 'write' function
that allows direct writing to the page. This makes future expansion
easier allowing such expansion to take place without the need of
further listener signature changes.