Reorganize the Node examples
The first four examples look fine and are getting a bit of moving
around to allow a more natural learning flow starting with example
number one.

What was example 4 (sendsms) is now example 1 (sms_sending). This
is done because that is the most straight-forward display of how
interaction with the API takes place and that deserves to be first.

What was example 1 (hours) is now example 2 (sms_replying) as it
is still pretty damn simple.

What was example 2 (groupsms) is now example 3 (sms_forwarding_groups)
as it mixes what are now examples 1 & 2, sending out text messages and
handling incoming messages in a bit more advanced fashion.

What was example 3 (textwall) is now example 4 (sms_wall) as it is a
bit more advanced than sms_forwarding_groups, involving data storage
instead of simple fire-and-forget logics. It is not really that hard
to wrap your head around and could be argued to be moved back to
number 3 later down the line, switching it out with group forwarding.