Deviate a bit from the original structure
Example 6 is now about handling incoming calls. That seems reasonable
right after example 5 which is about making outgoing calls. What was
example 7 prior to this (callnplay) was merged into this. To be fair,
I have not tested out whether the API takes in the JSON output from
these requests, but based on docs, it should. Probably. Will need to
test later.

Example 7 is not what was example 6 (whisper). That's a simple move
and nothing special is going on there.